Update StimulusSequence API.
When create a StimulusSequence object, a stimulus_id is now required
while a representation (z) is optional. This is in acknowledgement of
the fact that a user can always provide a stimulus_id, but may not be
able to provide a representation. In addition, the representation may
be part of the modeling process and may change.

Update default bounds to reflect assumptions of gradient equations.